rename manifest typings file
