+++ shared/manifest.ts
